+++ src/components/Logout/Logout.js
@@ -1,8 +1,6 @@
-import { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { useLocation } from 'react-router';
 import { FormattedMessage } from 'react-intl';
-import { useQueryClient } from 'react-query';
 
 import {
   Button,
@@ -15,67 +13,47 @@ import {
 import OrganizationLogo from '../OrganizationLogo';
 import { useStripes } from '../../StripesContext';
 import {
-  LOGOUT_TIMEOUT,
+  LOGOUT_MESSAGES,
   getUnauthorizedPathFromSession,
-  logout,
   removeUnauthorizedPathFromSession,
 } from '../../loginServices';
+import { useLogoutQuery } from './useLogoutQuery';
 
 import styles from './Logout.css';
 
 /**
  * Logout
- * Make an API call to /logout to terminate the session server-side and render
- * a "you've logged out" message. This component can be invoked from different
- * UI routes and will display a different message depending on that route:
- *
- * /logout: a user chose to end the session
- * /logout-timeout: stripes redirects to this location due to idle-session timeout
- *    the "reason" key in the query string provides additional details
+ * call useLogoutQuery() to terminate the session server-side, remove cookies,
+ * and clear all browser storage and react storage such as redux, react-query,
+ * and session timers.
  */
 const Logout = ({ sessionTimeoutTimer, sessionTimeoutWarningTimer }) => {
-  const { branding, okapi, store } = useStripes();
-  const [didLogout, setDidLogout] = useState(false);
+  const { branding, okapi } = useStripes();
   const location = useLocation();
-  const queryClient = useQueryClient();
+
+  // abuse useQuery to run once on-load, which AFAICT is impossible to do
+  // with useMutation in useEffect with a properly defined dependency array.
+  // that is, useMutation returns a function that it expects you to call in an
+  // event handler, rather than automatically on-load. so, here we are, calling
+  // a query and ignoring its response.
+  // filter out empty timers (they will be undefined when no session is active)
+  useLogoutQuery([sessionTimeoutTimer, sessionTimeoutWarningTimer].filter(Boolean));
 
   let messageId = null;
-  if (location.pathname === '/logout-timeout') {
-    const messages = {
-      [LOGOUT_TIMEOUT.ERROR]: 'stripes-core.rtr.error',
-      [LOGOUT_TIMEOUT.EXPIRED]: 'stripes-core.rtr.expired',
-      [LOGOUT_TIMEOUT.INACTIVITY]: 'stripes-core.rtr.idleSession.sessionExpiredSoSad',
-    };
+  const messages = {
+    [LOGOUT_MESSAGES.ERROR]: 'stripes-core.rtr.error',
+    [LOGOUT_MESSAGES.EXPIRED]: 'stripes-core.rtr.expired',
+    [LOGOUT_MESSAGES.INACTIVITY]: 'stripes-core.rtr.idleSession.sessionExpiredSoSad',
+    [LOGOUT_MESSAGES.INIT_ERROR]: 'stripes-core.init.error',
+  };
 
-    const params = new URLSearchParams(location.search);
-    messageId = messages[params.get('reason')] || messages[LOGOUT_TIMEOUT.INACTIVITY];
-  }
-
-
+  // pluck `reason` from the query string and use it to provide additional
+  // information about why this logout occurred
+  const params = new URLSearchParams(location.search);
+  messageId = messages[params.get('reason')];
   if (!messageId) {
     messageId = 'stripes-core.logoutComplete';
   }
-
-
-  useEffect(
-    () => {
-      if (okapi.isAuthenticated) {
-        sessionTimeoutTimer.clear();
-        sessionTimeoutWarningTimer.clear();
-
-        // returns a promise, which we ignore
-        logout(okapi.url, store, queryClient)
-          .then(setDidLogout(true));
-      } else {
-        setDidLogout(true);
-      }
-    },
-    // no dependencies because we only want to start the logout process once.
-    // we don't care about changes to stripes; certainly it'll be updated as
-    // part of the logout process
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
-  );
 
   const handleClick = (_e) => {
     removeUnauthorizedPathFromSession();
@@ -83,7 +61,7 @@ const Logout = ({ sessionTimeoutTimer, sessionTimeoutWarningTimer }) => {
 
   const redirectTo = getUnauthorizedPathFromSession() || '/';
 
-  if (!didLogout) {
+  if (okapi.isAuthenticated) {
     return <LoadingView />;
   }