+++ src/components/Logout/Logout.test.js
@@ -1,11 +1,17 @@
-import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
+import { act, render, screen, waitFor } from '@folio/jest-config-stripes/testing-library/react';
 import { userEvent } from '@folio/jest-config-stripes/testing-library/user-event';
 import { useLocation } from 'react-router';
 
 import Logout from './Logout';
+import { useLogoutQuery } from './useLogoutQuery';
 import { useStripes } from '../../StripesContext';
-import { getUnauthorizedPathFromSession, logout, setUnauthorizedPathToSession } from '../../loginServices';
+import {
+  LOGOUT_MESSAGES,
+  getUnauthorizedPathFromSession,
+  setUnauthorizedPathToSession
+} from '../../loginServices';
 
+jest.mock('./useLogoutQuery');
 jest.mock('../OrganizationLogo');
 jest.mock('../../StripesContext');
 jest.mock('react-router');
@@ -14,6 +20,7 @@ jest.mock('react-query');
 jest.mock('../../loginServices', () => ({
   ...jest.requireActual('../../loginServices'),
   logout: jest.fn(() => Promise.resolve()),
+  getLogoutTenant: jest.fn(() => ({ tenantId: 'tenant' })),
 }));
 
 const mockBranding = { branding: { logo: { src: './test.png' }, favicon: { src: './test-icon.png' } } };
@@ -33,16 +40,23 @@ describe('Logout', () => {
       screen.getByText('stripes-core.logoutComplete');
     });
 
-    it('if authenticated, calls logout then renders a logout message', async () => {
+    it('if authenticated, calls useLogoutQuery and renders loading dots ...', async () => {
       const mockUseStripes = useStripes;
       mockUseStripes.mockReturnValue({
         branding: mockBranding,
         okapi: { isAuthenticated: true },
       });
 
-      render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
-      expect(logout).toHaveBeenCalled();
-      screen.getByText('stripes-core.logoutComplete');
+      const mockUseLogoutQuery = useLogoutQuery;
+      mockUseLogoutQuery.mockReturnValue({});
+
+      act(() => {
+        render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
+      });
+      waitFor(() => {
+        expect(mockUseLogoutQuery).toHaveBeenCalled();
+        screen.getByText('LoadingView');
+      });
     });
 
     it('"log in again" href points to "/"', async () => {
@@ -58,56 +72,62 @@ describe('Logout', () => {
   describe('Timeout logout', () => {
     beforeEach(() => {
       const mockUseLocation = useLocation;
-      mockUseLocation.mockReturnValue({ pathname: '/logout-timeout' });
-    });
-
-    describe('if authenticated', () => {
-      it('calls logout then renders a timeout message', async () => {
-        const mockUseStripes = useStripes;
-        mockUseStripes.mockReturnValue({
-          branding: mockBranding,
-          okapi: { isAuthenticated: true },
-        });
-
-        render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
-        expect(logout).toHaveBeenCalled();
-        screen.getByText('stripes-core.rtr.idleSession.sessionExpiredSoSad');
-      });
-
-      it('"login in again" href points to pre-timeout location', () => {
-        const previousPath = '/monkey?bagel';
-        setUnauthorizedPathToSession(previousPath);
-
-        const mockUseStripes = useStripes;
-        mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false }, branding: mockBranding });
-
-        render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
-
-        expect(screen.getByRole('button')).toHaveAttribute('href', previousPath);
-      });
-
-      it('clicking "log in again" clears pre-timeout location from storage', async () => {
-        setUnauthorizedPathToSession('/monkey?bagel');
-
-        const user = userEvent.setup();
-        const mockUseStripes = useStripes;
-        mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false }, branding: mockBranding });
-
-        render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
-
-        await user.click(screen.getByRole('button'));
-        expect(getUnauthorizedPathFromSession()).toBeFalsy();
+      mockUseLocation.mockReturnValue({
+        pathname: '/logout',
+        search: `?reason=${LOGOUT_MESSAGES.INACTIVITY}`,
       });
     });
 
-    describe('if not authenticated', () => {
-      it('renders a timeout message', async () => {
-        const mockUseStripes = useStripes;
-        mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false }, branding: mockBranding });
+    it('if not authenticated, renders a timeout message', async () => {
+      const mockUseStripes = useStripes;
+      mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false }, branding: mockBranding });
 
-        render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
-        screen.getByText('stripes-core.rtr.idleSession.sessionExpiredSoSad');
+      render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
+      screen.getByText('stripes-core.rtr.idleSession.sessionExpiredSoSad');
+    });
+
+    it('if authenticated, calls useLogoutQuery and renders waiting dots ...', async () => {
+      const mockUseStripes = useStripes;
+      mockUseStripes.mockReturnValue({
+        branding: mockBranding,
+        okapi: { isAuthenticated: true },
       });
+
+      const mockUseLogoutQuery = useLogoutQuery;
+      mockUseLogoutQuery.mockReturnValue({});
+
+      act(() => {
+        render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
+      });
+      await waitFor(async () => {
+        expect(mockUseLogoutQuery).toHaveBeenCalled();
+        screen.getByText('LoadingView');
+      });
+    });
+
+    it('"login in again" href points to pre-timeout location', () => {
+      const previousPath = '/monkey?bagel';
+      setUnauthorizedPathToSession(previousPath);
+
+      const mockUseStripes = useStripes;
+      mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false }, branding: mockBranding });
+
+      render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
+
+      expect(screen.getByRole('button')).toHaveAttribute('href', previousPath);
+    });
+
+    it('clicking "log in again" clears pre-timeout location from storage', async () => {
+      setUnauthorizedPathToSession('/monkey?bagel');
+
+      const user = userEvent.setup();
+      const mockUseStripes = useStripes;
+      mockUseStripes.mockReturnValue({ okapi: { isAuthenticated: false }, branding: mockBranding });
+
+      render(<Logout sessionTimeoutTimer={{ clear: jest.fn() }} sessionTimeoutWarningTimer={{ clear: jest.fn() }} />);
+
+      await user.click(screen.getByRole('button'));
+      expect(getUnauthorizedPathFromSession()).toBeFalsy();
     });
   });
 });